new updates

+++ src/main/java/bookcrossing/controller/BookController.java
@@ -2,11 +2,7 @@ package bookcrossing.controller;
 
 import bookcrossing.domain.Book;
 import bookcrossing.service.BookService;
-import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
-import io.swagger.v3.oas.annotations.parameters.RequestBody;
-import io.swagger.v3.oas.annotations.responses.ApiResponse;
-import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -14,6 +10,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
@@ -40,14 +37,22 @@ public class BookController {
     @GetMapping("/{id}")
     public ResponseEntity<Book> getBookById(@PathVariable("id") Long id) {
         Optional<Book> book = bookService.getBookById(id);
-        return book.map(value -> new ResponseEntity<>(value, HttpStatus.OK)).orElseGet(() -> new ResponseEntity<>(HttpStatus.CONFLICT));
+        return book.map(value -> new ResponseEntity<>(value, HttpStatus.OK))
+                .orElseGet(() -> new ResponseEntity<>(HttpStatus.CONFLICT));
     }
-
 
     @GetMapping("/findByName")
     public ResponseEntity<List<Book>> findAllByName(@RequestParam("name") String name) {
-        List<Book> books = bookService.findAllByName(name);
-        return new ResponseEntity<>(books, HttpStatus.OK);
+        Optional<List<Book>> books = Optional.ofNullable(bookService.findAllByName(name));
+        return books.map(value -> new ResponseEntity<>(value, HttpStatus.OK))
+                .orElseGet(() -> new ResponseEntity<>(HttpStatus.CONFLICT));
+    }
+
+    @GetMapping("/available")
+    public ResponseEntity<List<Book>> getAllAvailableBooks() {
+        Optional<List<Book>> availableBooks = Optional.ofNullable(bookService.getAllAvailableBooks());
+        return availableBooks.map(value -> new ResponseEntity<>(value, HttpStatus.OK))
+                .orElseGet(() -> new ResponseEntity<>(HttpStatus.CONFLICT));
     }
 
     @PostMapping
@@ -55,28 +60,23 @@ public class BookController {
         return new ResponseEntity<>(bookService.createBook(book) ? HttpStatus.CREATED : HttpStatus.CONFLICT);
     }
 
-    @PutMapping //TODO: @Valid BindingResult
-    public ResponseEntity<HttpStatus> update(@RequestBody Book book) {
-        return new ResponseEntity<>(bookService.updateBook(book) ? HttpStatus.NO_CONTENT : HttpStatus.CONFLICT);
+    @PutMapping("/update/{id}")
+    public ResponseEntity<Book> updateBook(@PathVariable Long id, @RequestBody Book updatedBook) {
+        Optional<Book> updated = bookService.updateBook(id, updatedBook);
+
+        return updated.map(ResponseEntity::ok)
+                .orElseGet(() -> new ResponseEntity<>(HttpStatus.NOT_FOUND));
     }
 
-
-    @PutMapping("/{name}/{id}")
-    public ResponseEntity<HttpStatus> updateName(@PathVariable String name, @PathVariable Long id) {
-        bookService.updateName(name, id);
-        return new ResponseEntity<>(HttpStatus.OK);
-    }
-
-    @Operation(summary = "Удаляет пользователя из бд",
-            description = "Берет id, идет в бд и удаляет запись с этой id")
-    @ApiResponses(value = {
-            @ApiResponse(responseCode = "400", description = "Неверный ID"),
-            @ApiResponse(responseCode = "404", description = "id не найдено"),
-            @ApiResponse(responseCode = "204", description = "Значит что успешно удалили")})
     @DeleteMapping("/{id}")
     public ResponseEntity<HttpStatus> delete(@Parameter(description = "ID of the person to be deleted") @PathVariable("id") Long id) {
-        bookService.deleteBookById(id);
-        return new ResponseEntity<>(HttpStatus.NO_CONTENT);
-    }
+        Optional<Book> bookToDelete = bookService.deleteBookById(id);
 
+        if (bookToDelete.isPresent()) {
+            bookService.deleteBookById(id);
+            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        } else {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+    }
 }
